Unify paths more
