+++ packages/config/src/android/Paths.ts
@@ -31,3 +31,31 @@ export function getAndroidBuildGradle(projectRoot: string): string {
 export function getAppBuildGradle(projectRoot: string): string {
   return path.join(projectRoot, 'android', 'app', 'build.gradle');
 }
+
+export async function getAndroidManifestAsync(projectRoot: string): Promise<string> {
+  const shellPath = path.join(projectRoot, 'android');
+  if ((await fs.stat(shellPath)).isDirectory()) {
+    const manifestPath = path.join(shellPath, 'app/src/main/AndroidManifest.xml');
+    if ((await fs.stat(manifestPath)).isFile()) {
+      return manifestPath;
+    }
+  }
+
+  throw new Error(`Could not find AndroidManifest.xml in project directory: "${projectRoot}"`);
+}
+
+export async function getResourceXMLAsync(
+  projectDir: string,
+  { kind = 'values', name }: { kind?: string; name: string }
+): Promise<string> {
+  const shellPath = path.join(projectDir, 'android');
+  if ((await fs.stat(shellPath)).isDirectory()) {
+    const stylesPath = path.join(shellPath, `app/src/main/res/${kind}/${name}.xml`);
+    await fs.ensureFile(stylesPath);
+    return stylesPath;
+  }
+
+  throw new Error(
+    `Could not find android/app/src/main/res/${kind}/${name}.xml because the android project folder is missing.`
+  );
+}
