+++ packages/config/src/android/Strings.ts
@@ -2,7 +2,8 @@ import fs from 'fs-extra';
 import path from 'path';
 import { Builder } from 'xml2js';
 
-import { Document, getProjectXMLPathAsync, readXMLAsync } from './Manifest';
+import { Document, readXMLAsync } from './Manifest';
+import { getResourceXMLAsync } from './Paths';
 import { XMLItem } from './Styles';
 
 const BASE_STRINGS_XML = `<resources></resources>`;
@@ -10,8 +11,8 @@ const BASE_STRINGS_XML = `<resources></resources>`;
 export async function getProjectStringsXMLPathAsync(
   projectDir: string,
   { kind = 'values' }: { kind?: string } = {}
-): Promise<string | null> {
-  return getProjectXMLPathAsync(projectDir, { kind, name: 'strings' });
+): Promise<string> {
+  return getResourceXMLAsync(projectDir, { kind, name: 'strings' });
 }
 
 export async function readStringsXMLAsync(stringsPath: string): Promise<Document> {
